only tcpdump for specific scenario

+++ run_measurement.sh
@@ -54,11 +54,14 @@ coredns_path="../coredns"
 dnsproxy_path="../dnsproxy"
 chrome_path="../chromium"
 
-ip netns exec $namespace1 tcpdump -G 3600 -i any -w $root_dir/client-${timestamp}-${experiment_type}-${msmID}.pcap &
-tcpdumpclientPID=$!
-ip netns exec $namespace2 tcpdump -G 3600 -i any -w $root_dir/server-${timestamp}-${experiment_type}-${msmID}.pcap &
-tcpdumpserverPID=$!
-sleep 5
+if [[$website_under_test == "www.wikipedia.org"]]; then
+  echo "website is wikipedia, starting tcpdump to debug 0-rtt PLTs"
+  ip netns exec $namespace1 tcpdump -G 3600 -i any -w $root_dir/client-${timestamp}-${experiment_type}-${msmID}.pcap &
+  tcpdumpclientPID=$!
+  ip netns exec $namespace2 tcpdump -G 3600 -i any -w $root_dir/server-${timestamp}-${experiment_type}-${msmID}.pcap &
+  tcpdumpserverPID=$!
+  sleep 5
+fi
 
 cd $root_dir && cd $coredns_path
 #echo "starting coredns for DoQ udp:8853, DoUDP udp:53 and DoH tcp:443"
@@ -162,8 +165,10 @@ ip netns exec $namespace1 /home/quic_net01/.pyenv/shims/python3 chromium_measure
 
 kill -SIGTERM $corednsPID
 
-kill -SIGINT $tcpdumpclientPID
-kill -SIGINT $tcpdumpserverPID
+if [[$website_under_test == "www.wikipedia.org"]]; then
+  kill -SIGINT $tcpdumpclientPID
+  kill -SIGINT $tcpdumpserverPID
+fi
 # restart systemd-resolved
 #systemctl enable systemd-resolved
 #systemctl start systemd-resolved
